RC2

+++ MonoGame.Mathmatics.Test/Bounding.cs
@@ -5,6 +5,7 @@
 using System;
 using Microsoft.Xna.Framework;
 using NUnit.Framework;
+using Portland.Mathmatics;
 
 namespace MonoGame.Tests.Framework
 {
@@ -92,7 +93,7 @@ namespace MonoGame.Tests.Framework
         public void BoundingFrustumToBoundingBoxTests()
         {
             var view = Matrix.CreateLookAt(new Vector3(0, 0, 5), Vector3.Zero, Vector3.Up);
-            var projection = Matrix.CreatePerspectiveFieldOfView(MathHelper.PiOver4, 1, 1, 100);
+            var projection = Matrix.CreatePerspectiveFieldOfView((float)MathHelper.PiOver4, 1, 1, 100);
             var testFrustum = new BoundingFrustum(view * projection);
 
             var bbox1 = new BoundingBox(new Vector3(0, 0, 0), new Vector3(1, 1, 1));
@@ -116,7 +117,7 @@ namespace MonoGame.Tests.Framework
         public void BoundingFrustumToBoundingFrustumTests()
         {
             var view = Matrix.CreateLookAt(new Vector3(0, 0, 5), Vector3.Zero, Vector3.Up);
-            var projection = Matrix.CreatePerspectiveFieldOfView(MathHelper.PiOver4, 1, 1, 100);
+            var projection = Matrix.CreatePerspectiveFieldOfView((float)MathHelper.PiOver4, 1, 1, 100);
             var testFrustum = new BoundingFrustum(view * projection);
 
             // Same frustum.
@@ -127,7 +128,7 @@ namespace MonoGame.Tests.Framework
 
             // Smaller frustum contained entirely inside.
             var view2 = Matrix.CreateLookAt(new Vector3(0, 0, 4), Vector3.Zero, Vector3.Up);
-            var projection2 = Matrix.CreatePerspectiveFieldOfView(MathHelper.PiOver4, 1, 1, 50);
+            var projection2 = Matrix.CreatePerspectiveFieldOfView((float)MathHelper.PiOver4, 1, 1, 50);
             otherFrustum.Matrix = view2 * projection2;
 
             Assert.That(testFrustum.Contains(otherFrustum), Is.EqualTo(ContainmentType.Contains));
@@ -148,7 +149,7 @@ namespace MonoGame.Tests.Framework
 
             // Larger frustum, entirely containing test frustum.
             var view4 = Matrix.CreateLookAt(new Vector3(0, 0, 10), Vector3.Zero, Vector3.Up);
-            var projection4 = Matrix.CreatePerspectiveFieldOfView(MathHelper.PiOver4, 1, 1, 1000);
+            var projection4 = Matrix.CreatePerspectiveFieldOfView((float)MathHelper.PiOver4, 1, 1, 1000);
             otherFrustum.Matrix = view4 * projection4;
 
             Assert.That(testFrustum.Contains(otherFrustum), Is.EqualTo(ContainmentType.Intersects));
@@ -156,7 +157,7 @@ namespace MonoGame.Tests.Framework
 
             var bf =
                 new BoundingFrustum(Matrix.CreateLookAt(new Vector3(0, 1, 1), new Vector3(0, 0, 0), Vector3.Up) *
-                                    Matrix.CreatePerspectiveFieldOfView(MathHelper.PiOver4,
+                                    Matrix.CreatePerspectiveFieldOfView((float)MathHelper.PiOver4,
                                         1.3f, 0.1f, 1000.0f));
             var ray = new Ray(new Vector3(0, 0.5f, 0.5f), new Vector3(0, 0, 0));
             var ray2 = new Ray(new Vector3(0, 1.0f, 1.0f), new Vector3(0, 0, 0));
